requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==3.2
django-auditlog==2.1.1
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
mysqlclient==2.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
